source/lambda-functions/image-processor/requirements.txt
Pillow==10.0.0
boto3==1.28.85

